chore(clerk-js): Remove barrel file
